# src/main/java/org/andrejstrogonov/moskowstockbackend/dto/EarlyRepayment.java
package org.andrejstrogonov.moskowstockbackend.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class EarlyRepayment {

    private Integer month;
    private Double amount;
}